finish

+++ src/main/kotlin/com/example/demo/controllers/NoteController.kt
@@ -3,7 +3,10 @@ package com.example.demo.controllers
 import com.example.demo.controllers.NoteController.NoteResponse
 import com.example.demo.database.model.Note
 import com.example.demo.database.model.repository.NoteRepository
+import jakarta.validation.Valid
+import jakarta.validation.constraints.NotBlank
 import org.bson.types.ObjectId
+import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
@@ -17,10 +20,12 @@ import java.time.Instant
 @RestController
 @RequestMapping("/notes")
 class NoteController (
-    private val repository: NoteRepository
+    private val repository: NoteRepository,
+    private val noteRepository: NoteRepository
 ){
 
     data class NoteRequest (
+        @field:NotBlank(message = "Title is cant be blank")
         val title: String,
         val content:String,
         val color:Long,
@@ -37,7 +42,8 @@ class NoteController (
 
     @PostMapping
     fun save(
-        @RequestBody body: NoteRequest ) : NoteResponse{
+        @Valid @RequestBody body: NoteRequest ) : NoteResponse{
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
         val note = repository.save(
             Note(
                 id = body.id?.let { ObjectId(it)} ?: ObjectId.get(),
@@ -45,7 +51,7 @@ class NoteController (
                 content = body.content,
                 color = body.color,
                 createdAt = Instant.now(),
-                ownerId = ObjectId()
+                ownerId = ObjectId(ownerId)
 
             )
         )
@@ -53,9 +59,8 @@ class NoteController (
     }
 
     @GetMapping
-    fun findbyOwnerId(
-        @RequestParam(required = true) ownerId: String
-    ) : List<NoteResponse> {
+    fun findbyOwnerId() : List<NoteResponse> {
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
         return repository.findByOwnerId(ObjectId(ownerId)).map {
             it.toResponse()
         }
@@ -63,7 +68,13 @@ class NoteController (
 
     @DeleteMapping(path = ["/{id}"])
     fun deleteById(@PathVariable id: String){
-        repository.deleteById(ObjectId(id))
+        val note = noteRepository.findById(ObjectId(id)).orElseThrow{
+            IllegalArgumentException("Note not Found")
+        }
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
+        if (note.ownerId.toHexString() == ownerId) {
+            repository.deleteById(ObjectId(id))
+        }
     }
 }
